Rename github_status to github.
