+++ bin/gh_status
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
 
-import circleci.github_status
+import circleci.github
 
-circleci.github_status.cli()
+circleci.github.status_cli()
